% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy}
\alias{copy}
\title{Copy}
\usage{
copy(source, target, .warn = FALSE)
}
\arguments{
\item{source}{source directory.}

\item{target}{target directory.}

\item{.warn}{if \code{TRUE}, warns if file from source is already in target.}
}
\value{
Character vector with files already in target.
}
\description{
Copies all files in source to folder target and returns filenames that were
already in target.
}
\keyword{internal}
