% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{scaffold}
\alias{scaffold}
\title{Scaffold a Paper Vignette}
\usage{
scaffold(template)
}
\arguments{
\item{template}{Template to use (needs to match one of the templates available in \code{inst/templates}).}
}
\value{
Returns 1 invisibly if all went well.
}
\description{
Scaffolds a vignette paper using the JSS template https://www.jstatsoft.org/style.
}
